refactor: img 태그 오류 수정
