+++ src/pages/mainPage/MainPage.tsx
@@ -2,8 +2,8 @@ import React from "react";
 import LandingScreen from "./LandingScreen/LandingScreen";
 import NavigationBar from "./NavigationBar/NavigationBar";
 import MainScreen from "./MainScreen/MainScreen";
-import NotificationScreen from "./NotificationScreen";
-import InfoScreen from "./InfoScreen";
+import NotificationScreen from "./NotificationScreen/NotificationScreen";
+import InfoScreen from "./InfoScreen/InfoScreen";
 
 const MainPage = () => {
   return (
