+++ src/pages/mainPage/MainScreen/FCFSEventSection.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Timer from "../Timer";
+import Timer from "./Timer";
 import EventHeader from "../../../components/mainPage/MainScreen/EventHeader";
 import EventFooter from "../../../components/mainPage/MainScreen/EventFooter";
 
@@ -8,7 +8,7 @@ const FCFSEventSection = () => {
     <>
       <EventHeader title="event" description="desc" />
       <Timer />
-      <img>coffee img</img>
+      <img />
       <EventFooter
         eventInfo={{ title: "title", description: "des", option: Object }}
         buttonInfo={{ size: "small", onClick: () => {}, isEnabled: true }}
